decapitalized tables
For consistency and so this isn't a pain to maintain, I'm decapitalizing all table names. I think it might not matter much in our actual queries because MySQL isn't case sensitive, but we should look at a standard for whether columns should be capitalized or not, because that can actually make a difference I think.